move AI class initialization to config
